requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
redis
pytest
pytest-django
drf-yasg